Cleanup lib
* Removed unneeded files
* move config and vintage_net_gadget.ex to root level dir
* cleanup mix.exs
