+++ .dialyzer_ignore.exs
@@ -2,5 +2,4 @@
 #
 # See https://github.com/jeremyjh/dialyxir#elixir-term-format
 [
-  {"lib/vintage_net/interface/classification.ex", :contract_supertype}
 ]
