+++ mix.exs
@@ -9,7 +9,8 @@ defmodule VintageNetGadget.MixProject do
       app: :vintage_net_gadget,
       version: @version,
       elixir: "~> 1.7",
-      elixirc_paths: elixirc_paths(Mix.env()),
+      elixirc_paths: ["vintage_net_gadget.ex"],
+      config_path: "config.exs",
       test_coverage: [tool: ExCoveralls],
       start_permanent: Mix.env() == :prod,
       build_embedded: true,
@@ -27,9 +28,6 @@ defmodule VintageNetGadget.MixProject do
     ]
   end
 
-  defp elixirc_paths(:test), do: ["lib", "test/support"]
-  defp elixirc_paths(_), do: ["lib"]
-
   defp description do
     "USB Gadget networking for VintageNet"
   end
@@ -37,7 +35,7 @@ defmodule VintageNetGadget.MixProject do
   defp package do
     %{
       files: [
-        "lib",
+        "vintage_net_gadget.ex",
         "test",
         "mix.exs",
         "README.md",
@@ -54,7 +52,8 @@ defmodule VintageNetGadget.MixProject do
       {:ex_doc, "~> 0.19", only: :docs, runtime: false},
       {:excoveralls, "~> 0.8", only: :test, runtime: false},
       {:dialyxir, "~> 1.0.0-rc.6", only: [:dev, :test], runtime: false},
-      {:vintage_net, "~> 0.7.0", path: "../vintage_net", override: true},
+      {:vintage_net,
+       github: "nerves-networking/vintage_net", branch: "the-split", override: true},
       {:one_dhcpd, "~> 0.2.3"}
     ]
   end
@@ -62,7 +61,6 @@ defmodule VintageNetGadget.MixProject do
   defp dialyzer() do
     [
       flags: [:race_conditions, :unmatched_returns, :error_handling, :underspecs],
-      plt_add_apps: [:busybox],
       ignore_warnings: ".dialyzer_ignore.exs",
       list_unused_filters: true
     ]
