+++ test/vintage_net_gadget_test.exs
@@ -1,28 +1,26 @@
-defmodule VintageNet.Technology.GadgetTest do
+defmodule VintageNetGadgetTest do
   use ExUnit.Case
   alias VintageNet.Interface.RawConfig
-  alias VintageNet.Technology.Gadget
-
-  import VintageNetTest.Utils
+  alias VintageNetGadget, as: Gadget
 
   test "normalization simplifies configuration" do
-    input = %{type: VintageNet.Technology.Gadget, random_field: 42}
+    input = %{type: Gadget, random_field: 42}
 
-    assert Gadget.normalize(input) == %{type: VintageNet.Technology.Gadget, gadget: %{}}
+    assert Gadget.normalize(input) == %{type: Gadget, gadget: %{}}
   end
 
   test "normalization preserves hostname override" do
-    input = %{type: VintageNet.Technology.Gadget, gadget: %{hostname: "my_host"}}
+    input = %{type: Gadget, gadget: %{hostname: "my_host"}}
 
     assert Gadget.normalize(input) == %{
-             type: VintageNet.Technology.Gadget,
+             type: Gadget,
              gadget: %{hostname: "my_host"}
            }
   end
 
   test "create a gadget configuration" do
     input = %{
-      type: VintageNet.Technology.Gadget,
+      type: Gadget,
       gadget: %{hostname: "test_hostname"}
     }
 
@@ -30,7 +28,7 @@ defmodule VintageNet.Technology.GadgetTest do
 
     expected = %RawConfig{
       ifname: "usb0",
-      type: VintageNet.Technology.Gadget,
+      type: Gadget,
       source_config: input,
       child_specs: [
         %{id: {OneDHCPD, "usb0"}, start: {OneDHCPD, :start_server, ["usb0"]}},
@@ -54,5 +52,23 @@ defmodule VintageNet.Technology.GadgetTest do
     }
 
     assert expected == output
+  end
+
+  def default_opts() do
+    # Use the defaults in mix.exs, but normalize the paths to commands
+    Application.get_all_env(:vintage_net)
+    |> Keyword.merge(
+      bin_chat: "chat",
+      bin_dnsd: "dnsd",
+      bin_ifdown: "ifdown",
+      bin_ifup: "ifup",
+      bin_ip: "ip",
+      bin_killall: "killall",
+      bin_mknod: "mknod",
+      bin_pppd: "pppd",
+      bin_udhcpc: "udhcpc",
+      bin_udhcpd: "udhcpd",
+      bin_wpa_supplicant: "wpa_supplicant"
+    )
   end
 end
